requirements.txt
python-telegram-bot
qrcode
Pillow
aiohttp
bs4
requests